# api/openapi.yaml
openapi: 3.0.3
info:
  title: devm Control Plane API
  version: 0.1.0
  description: Public API used by the devm CLI and web application.
servers:
  - url: /v1
security:
  - workosBearer: []
paths:
...
  /capsules/{name}/tags/{tag}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$'
      - name: tag
        in: path
        required: true
        schema:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$'
    get:
      summary: Resolve an owned Capsule tag
      operationId: getCapsuleTag
      responses:
        '200':
          description: The current owner-scoped Capsule tag pointer.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestID' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CapsuleTag' }
        default:
          $ref: '#/components/responses/Error'
    put:
      summary: Publish an owned Capsule tag
      operationId: putCapsuleTag
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PutCapsuleTagRequest' }
      responses:
        '200':
          description: The current owner-scoped Capsule tag pointer. Publishing the same digest is idempotent; publishing a different digest retags it.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestID' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CapsuleTag' }
        default:
          $ref: '#/components/responses/Error'
...
  /project-seeds:
    post:
      summary: Register a Project Seed
      operationId: createProjectSeed
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectSeedInput' }
      responses:
        '201':
          description: Immutable Project Seed registered.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestID' }
          content:
            application/json:
              schema:
                type: object
                required: [id, digest]
                properties:
                  id: { type: string }
                  digest: { type: string }
        default:
          $ref: '#/components/responses/Error'
  /environments:
    get:
      summary: List Environments
      operationId: listEnvironments
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Environments owned by the user.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestID' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnvironmentPage' }
        default:
          $ref: '#/components/responses/Error'
    post:
      summary: Create an Environment
      operationId: createEnvironment
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEnvironmentInput' }
      responses:
        '202':
          description: Environment creation accepted.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestID' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnvironmentOperation' }
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    workosBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 512 }
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  headers:
    RequestID:
      description: Stable identifier for this API request.
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Product-level error.
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ProjectSeedInput:
      type: object
      required: [repositoryUrl, baseRevision, digest, manifestDigest]
      properties:
        repositoryUrl: { type: string, format: uri }
        baseRevision: { type: string }
        digest: { type: string, pattern: '^sha256:[a-f0-9]{64}$' }
        gitBundleDigest: { type: string, nullable: true, pattern: '^sha256:[a-f0-9]{64}$' }
        trackedPatchDigest: { type: string, nullable: true, pattern: '^sha256:[a-f0-9]{64}$' }
        untrackedBundleDigest: { type: string, nullable: true, pattern: '^sha256:[a-f0-9]{64}$' }
        manifestDigest: { type: string, pattern: '^sha256:[a-f0-9]{64}$' }
    AutoStopPolicy:
      type: object
      required: [mode, gracePeriodSeconds]
      properties:
        mode:
          type: string
          enum: [when_disconnected, when_agents_finish, when_fully_idle, manual]
        gracePeriodSeconds: { type: integer, minimum: 0, maximum: 86400 }
    CreateEnvironmentInput:
      type: object
      required: [name, region, runtimePreset, profileVersionId, projectSeedId, autoStopPolicy, sshKeyIds]
      properties:
        name: { type: string, minLength: 3, maxLength: 64 }
        region: { type: string }
        runtimePreset: { type: string }
        profileVersionId: { type: string }
        projectSeedId: { type: string }
        autoStopPolicy: { $ref: '#/components/schemas/AutoStopPolicy' }
        sshKeyIds:
          type: array
          minItems: 1
          items: { type: string }
    Runtime:
      type: object
      required: [id, status, runtimePreset, region, availabilityZone, imageVersion]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [absent, provisioning, starting, ready, stopping, stopped, replacing, error]
        runtimePreset: { type: string }
        region: { type: string }
        availabilityZone: { type: string }
        imageVersion: { type: string }
    Environment:
      type: object
      required: [id, name, slug, lifecycle, health, region, runtimePreset, pinnedProfileVersionId, capsuleLockId, autoStopPolicy, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        lifecycle: { type: string, enum: [creating, active, deleting, deleted] }
        health: { type: string, enum: [healthy, degraded, blocked, unknown] }
        region: { type: string }
        runtimePreset: { type: string }
        pinnedProfileVersionId: { type: string }
        capsuleLockId: { type: string, nullable: true }
        capsuleLock: { $ref: '#/components/schemas/CapsuleLock' }
        autoStopPolicy: { $ref: '#/components/schemas/AutoStopPolicy' }
        runtime: { $ref: '#/components/schemas/Runtime' }
        activeOperationId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    LockedCapsule:
      type: object
      required: [ref, digest]
      properties:
        ref: { type: string }
        digest: { type: string, pattern: '^sha256:[a-f0-9]{64}$' }
    CapsuleLock:
      type: object
      required: [id, digest, profileVersionId, projectCapsuleDigest, capsules]
      properties:
        id: { type: string }
        digest: { type: string, pattern: '^sha256:[a-f0-9]{64}$' }
        profileVersionId: { type: string }
        projectCapsuleDigest: { type: string, pattern: '^sha256:[a-f0-9]{64}$' }
        capsules:
          type: array
          items: { $ref: '#/components/schemas/LockedCapsule' }
...
    PutCapsuleTagRequest:
      type: object
      required: [digest]
      properties:
        digest: { type: string, pattern: '^sha256:[a-f0-9]{64}$' }
    CapsuleTag:
      type: object
      required: [name, tag, digest, updatedAt]
      properties:
        name: { type: string }
        tag: { type: string }
        digest: { type: string, pattern: '^sha256:[a-f0-9]{64}$' }
        updatedAt: { type: string, format: date-time }
    EnvironmentOperation:
      type: object
      required: [environment, operation]
      properties:
        environment: { $ref: '#/components/schemas/Environment' }
        operation: { $ref: '#/components/schemas/Operation' }
    EnvironmentPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Environment' }
        nextCursor: { type: string, nullable: true }
...
    OperationStep:
      type: object
      required: [stepKey, status, summary]
      properties:
        stepKey: { type: string }
        status: { type: string, enum: [pending, running, succeeded, failed, skipped, blocked] }
        summary: { type: string }
    Operation:
      type: object
      required: [id, environmentId, type, status, steps, createdAt]
      properties:
        id: { type: string }
        environmentId: { type: string }
        type: { type: string }
        status: { type: string, enum: [queued, running, succeeded, failed, cancelled, blocked] }
        steps:
          type: array
          items: { $ref: '#/components/schemas/OperationStep' }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
...
    ErrorResponse:
      type: object
      required: [requestId, error]
      properties:
        requestId: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            operationId: { type: string, nullable: true }
            safeState: { type: string, nullable: true }
            nextAction: { type: string, nullable: true }