# libs/db/internal/dbsql/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbsql

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type ActivitySnapshot struct {
	RuntimeID            string
	Sequence             int64
	EnvironmentID        string
	ObservedAt           pgtype.Timestamptz
	SshConnections       int32
	IdeConnections       int32
	CodexProcesses       int32
	ClaudeProcesses      int32
	ProtectedProcesses   int32
	SelectedContainers   int32
	UnknownUserProcesses int32
}

type CapsuleLock struct {
	ID                   string
	EnvironmentID        string
	ProfileVersionID     string
	ProjectCapsuleDigest string
	Digest               string
	Capsules             []byte
	ResolvedComponents   []byte
	CreatedAt            pgtype.Timestamptz
}

type CapsuleTag struct {
	OwnerUserID string
	Name        string
	Tag         string
	Digest      string
	UpdatedAt   pgtype.Timestamptz
}

type ComputeUsageInterval struct {
	ID                  string
	UserID              string
	EnvironmentID       string
	RuntimeID           string
	Region              string
	RuntimePreset       string
	StartedAt           pgtype.Timestamptz
	EndedAt             pgtype.Timestamptz
	ClosureSource       *string
	CreditTransactionID *string
}

type CreditBalance struct {
	UserID    string
	Credits   int64
	Version   int64
	UpdatedAt pgtype.Timestamptz
}

type CreditRate struct {
	Version        string
	ResourceType   string
	Region         string
	RuntimePreset  *string
	RawUnit        string
	CreditsPerUnit string
	EffectiveAt    pgtype.Timestamptz
}

type CreditTransaction struct {
	ID             string
	UserID         string
	Kind           string
	Credits        int64
	ResourceType   *string
	EnvironmentID  *string
	ResourceID     *string
	Region         *string
	RawQuantity    *string
	RawUnit        *string
	RateVersion    *string
	IdempotencyKey string
	OccurredAt     pgtype.Timestamptz
	CreatedAt      pgtype.Timestamptz
}

type EnvironmentMaterialization struct {
	EnvironmentID               string
	LockID                      string
	ID                          *string
	LockDigest                  string
	CapsuleDigest               string
	ComponentID                 string
	ComponentDigest             string
	AdapterID                   string
	AdapterVersion              string
	TargetAgentVersion          string
	Scope                       string
	ComponentType               string
	TrustClass                  string
	NonSecretOverridesDigest    *string
	SecretVersionIdentifiers    []byte
	EffectiveCacheKey           string
	Mode                        *string
	Root                        *string
	Target                      *string
	Selector                    *string
	Directory                   bool
	FilePaths                   []byte
	LastAppliedDigest           *string
	ObservedDigest              *string
	CredentialRequirementDigest *string
	CreatedAt                   pgtype.Timestamptz
	UpdatedAt                   pgtype.Timestamptz
}

type PolarCustomer struct {
	UserID          string
	PolarCustomerID string
	ExternalEventID string
	ObservedAt      pgtype.Timestamptz
}

type PolarDelivery struct {
	ExternalID          string
	CreditTransactionID string
	EventPayload        []byte
	CreatedAt           pgtype.Timestamptz
	DeliveredAt         pgtype.Timestamptz
}

type PolarWebhookReceipt struct {
	ExternalEventID string
	EventType       string
	PayloadSha256   []byte
	OccurredAt      pgtype.Timestamptz
	ReceivedAt      pgtype.Timestamptz
}

type Profile struct {
	ID          string
	OwnerUserID string
	Name        string
	Slug        string
	CreatedAt   pgtype.Timestamptz
	ArchivedAt  pgtype.Timestamptz
}

type ProfileVersion struct {
	ID              string
	ProfileID       string
	ParentVersionID *string
	Version         int64
	Digest          string
	CreatedAt       pgtype.Timestamptz
}

type SshKey struct {
	ID          string
	OwnerUserID string
	Label       string
	Algorithm   string
	Fingerprint string
	PublicKey   string
	CreatedAt   pgtype.Timestamptz
	RevokedAt   pgtype.Timestamptz
}

type Subscription struct {
	UserID              string
	PolarSubscriptionID string
	PolarCustomerID     string
	Status              string
	CurrentPeriodStart  pgtype.Timestamptz
	CurrentPeriodEnd    pgtype.Timestamptz
	CancelAtPeriodEnd   bool
	CanceledAt          pgtype.Timestamptz
	ExternalEventID     string
	ObservedAt          pgtype.Timestamptz
}

type UploadIntent struct {
	ID          string
	OwnerUserID string
	Kind        string
	Digest      string
	SizeBytes   int64
	ObjectKey   string
	CreatedAt   pgtype.Timestamptz
	ExpiresAt   pgtype.Timestamptz
}

type User struct {
	ID            string
	WorkosUserID  string
	DefaultRegion string
	CreatedAt     pgtype.Timestamptz
	UpdatedAt     pgtype.Timestamptz
}